fixed some clippy warnings
